+++ projects/swimlane/ngx-ui/src/lib/components/section/section-appearance.enum.ts
@@ -1,5 +1,6 @@
 export enum SectionAppearance {
   Legacy = 'legacy',
   Outline = 'outline',
-  Light = 'light'
+  Light = 'light',
+  Minimal = 'minimal'
 }