+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## HEAD (unreleased)
 
+- Feature (`ngx-date-range-picker`): Added a date range component
+
 ## 50.0.0-alpha.1 (2025-07-03)
 
 - Breaking: Renamed ngx-filter-select to ngx-filter