+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## HEAD (unreleased)
 
 - Feature (`ngx-date-range-picker`): Added a date range component
+- Enhancement (`ngx-filter`): Updated output click event to handle icon click
+- Enhancement (`ngx-filter`): Explicitly sync dropdownComponent.open for CustomDropdown type
 
 ## 50.0.0-alpha.1 (2025-07-03)