+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## HEAD (unreleased)
 
+## 50.0.0-alpha.2 (2025-07-11)
+
 - Feature (`ngx-date-range-picker`): Added a date range component
 - Enhancement (`ngx-filter`): Updated output click event to handle icon click
 - Enhancement (`ngx-filter`): Explicitly sync dropdownComponent.open for CustomDropdown type